requirements/base.txt
asgiref
Django
django-admin-volt
django-cors-headers
django-filter
djangorestframework
exceptiongroup
httplib2
iniconfig
packaging
Pillow
plantuml
pluggy
pyparsing
pytest
pytest-django
pytz
six
sqlparse
tomli
typing_extensions